pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foresight-x"
version = "0.1.0"
description = "Evidence-grounded decision agent (Foresight-X)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.2,<3",
    "llama-index>=0.10,<0.13",
    "llama-index-vector-stores-chroma>=0.1,<0.5",
    "llama-index-llms-openai>=0.1,<0.6",
    "llama-index-embeddings-openai>=0.1,<0.6",
    "chromadb>=0.4,<1",
    "tavily-python>=0.3,<0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "python-dotenv>=1,<2",
]

[tool.hatch.build.targets.wheel]
packages = ["foresight_x"]

[tool.hatch.build.targets.wheel.force-include]
"foresight_x/retrieval/seeds" = "foresight_x/retrieval/seeds"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 100
target-version = "py311"
